requirements.txt
pandas
matplotlib
mysql-connector-python